change body attribute to a body block for better tooling
This commit changes the body attribute to a body block for resources and composite status/ connection.
It renames the body attribute to a result attribute for functions.
This allows for better tooling and schema inference based on the apiVersion and kind attributes of the body block.

There is one regression, however. It is now impossible to write a function that returns the entire k8s object
since body is a block and no longer an attribute that can be assigned to.

In this commit, both old and new forms are accepted and the previous forms are merely deprecated.

Signed-off-by: gotwarlost <krishnan.anantheswaran@elastic.co>